+++ zetaclient/config/config_chain.go
@@ -22,7 +22,7 @@ func New(setDefaults bool) Config {
 	}
 
 	if setDefaults {
-		cfg.MempoolCongestionTxCount = DefaultMempoolCongestionTxCount
+		cfg.MempoolCongestionThreshold = DefaultMempoolCongestionThreshold
 		cfg.EVMChainConfigs = evmChainsConfigs()
 		cfg.BTCChainConfigs = btcChainsConfigs()
 		cfg.SolanaConfig = solanaConfigLocalnet()