+++ zetaclient/config/types_test.go
@@ -16,16 +16,16 @@ func Test_GetRelayerKeyPath(t *testing.T) {
 	require.Equal(t, config.DefaultRelayerKeyPath, cfg.GetRelayerKeyPath())
 }
 
-func Test_GetMempoolCongestionTxCount(t *testing.T) {
-	t.Run("should return default mempool congestion tx count", func(t *testing.T) {
+func Test_GetMempoolCongestionThreshold(t *testing.T) {
+	t.Run("should return default mempool congestion threshold", func(t *testing.T) {
 		cfg := config.New(false)
-		require.EqualValues(t, config.DefaultMempoolCongestionTxCount, cfg.GetMempoolCongestionTxCount())
+		require.EqualValues(t, config.DefaultMempoolCongestionThreshold, cfg.GetMempoolCongestionThreshold())
 	})
 
-	t.Run("should return configured mempool congestion tx count", func(t *testing.T) {
+	t.Run("should return configured mempool congestion threshold", func(t *testing.T) {
 		cfg := config.New(false)
-		cfg.MempoolCongestionTxCount = 5000
-		require.EqualValues(t, 5000, cfg.GetMempoolCongestionTxCount())
+		cfg.MempoolCongestionThreshold = 5000
+		require.EqualValues(t, 5000, cfg.GetMempoolCongestionThreshold())
 	})
 }